# aggregator: improvements to the deduplication algorithm