… pkg/aggregator/aggregator_benchmark_test.go