### What does this PR do?

Minor improvements to the deduplication logic in the aggregator. 
Instead of calling `sort.Strings` for small tag sets use a simple insertion sort.

### Motivation

Remove one allocation per ingested metric sample for small tag sets. The allocation is due to how `sort.Strings` works internally.

### Benchmarks

```
benchmark                          old ns/op     new ns/op     delta
BenchmarkDeduplicateTags2-4        71.5          23.9          -66.57%
BenchmarkDeduplicateTags4-4        99.5          68.2          -31.46%
BenchmarkDeduplicateTags10-4       238           223           -6.30%
BenchmarkDeduplicateTags20-4       636           656           +3.14%
BenchmarkDeduplicateTags40-4       1453          1455          +0.14%
BenchmarkDeduplicateTags60-4       2531          2534          +0.12%
BenchmarkDeduplicateTags100-4      4547          4724          +3.89%
BenchmarkDeduplicateTags1000-4     75130         74942         -0.25%

benchmark                          old allocs     new allocs     delta
BenchmarkDeduplicateTags2-4        1              0              -100.00%
BenchmarkDeduplicateTags4-4        1              0              -100.00%
BenchmarkDeduplicateTags10-4       1              0              -100.00%
BenchmarkDeduplicateTags20-4       1              0              -100.00%
BenchmarkDeduplicateTags40-4       1              1              +0.00%
BenchmarkDeduplicateTags60-4       1              1              +0.00%
BenchmarkDeduplicateTags100-4      1              1              +0.00%
BenchmarkDeduplicateTags1000-4     1              1              +0.00%

benchmark                          old bytes     new bytes     delta
BenchmarkDeduplicateTags2-4        32            0             -100.00%
BenchmarkDeduplicateTags4-4        32            0             -100.00%
BenchmarkDeduplicateTags10-4       32            0             -100.00%
BenchmarkDeduplicateTags20-4       32            0             -100.00%
BenchmarkDeduplicateTags40-4       32            32            +0.00%
BenchmarkDeduplicateTags60-4       32            32            +0.00%
BenchmarkDeduplicateTags100-4      32            32            +0.00%
BenchmarkDeduplicateTags1000-4     32            32            +0.00%
```